backend/requirements.txt
# ============================================
# DJANGO CORE
# ============================================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5

# ============================================
# DATABASE
# ============================================
psycopg2-binary==2.9.9

# ============================================
# AUTHENTICATION
# ============================================
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
PyJWT==2.8.0

# ============================================
# FILE HANDLING
# ============================================
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7

# ============================================
# ADMIN INTERFACE
# ============================================
django-admin-interface==0.26.0
django-nested-admin==4.0.2
django-json-widget==1.1.1
django-colorfield==0.11.0

# ============================================
# CELERY & TASK QUEUE
# ============================================
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ============================================
# AI SERVICES
# ============================================
openai==1.3.8
anthropic==0.8.0

# ============================================
# EXTERNAL INTEGRATIONS
# ============================================
CloudFlare==2.11.7
GitPython==3.1.40
requests==2.31.0
httpx==0.25.2

# ============================================
# WEB SERVER
# ============================================
gunicorn==21.2.0
whitenoise==6.6.0

# ============================================
# DEVELOPMENT TOOLS
# ============================================
ipython==8.18.1
django-extensions==3.2.3

# ============================================
# MONITORING
# ============================================
sentry-sdk==1.38.0

# ============================================
# UTILITIES
# ============================================
python-decouple==3.8
python-dateutil==2.8.2
python-slugify==8.0.1

# ============================================
# TESTING (Optional)
# ============================================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# ============================================
# CODE QUALITY (Optional)
# ============================================
flake8==6.1.0
black==23.12.0
isort==5.13.2
